각 핸들러 기능별로 분할
