+++ internal/handlers/students.go
@@ -1,0 +1,160 @@
+package handlers
+
+import (
+	"database/sql"
+	"encoding/json"
+	"log"
+	"net/http"
+
+	"attendance-management/internal/models"
+)
+
+func (h *Handler) StudentHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case http.MethodGet:
+		h.GetStudents(w, r)
+	case http.MethodPost:
+		h.CreateStudent(w, r)
+	default:
+		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
+	}
+}
+
+func (h *Handler) StudentByIDHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case http.MethodGet:
+		h.GetStudentByID(w, r)
+	case http.MethodPut:
+		h.UpdateStudentByID(w, r)
+	case http.MethodDelete:
+		h.DeleteStudentByID(w, r)
+	default:
+		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
+	}
+}
+
+func (h *Handler) GetStudents(w http.ResponseWriter, r *http.Request) {
+	rows, err := h.db.Query("SELECT * FROM students")
+	if err != nil {
+		log.Println("DB 조회 오류:", err)
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+		return
+	}
+	defer rows.Close()
+
+	var students []models.Student
+	for rows.Next() {
+		var s models.Student
+		if err := rows.Scan(&s.StudentID, &s.Name, &s.Grade, &s.Phone, &s.ParentPhone); err != nil {
+			log.Println("rows.Scan 오류:", err)
+			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+			return
+		}
+		students = append(students, s)
+	}
+
+	writeJSON(w, http.StatusOK, students)
+}
+
+func (h *Handler) CreateStudent(w http.ResponseWriter, r *http.Request) {
+	var student models.Student
+	if err := json.NewDecoder(r.Body).Decode(&student); err != nil {
+		http.Error(w, "Invalid JSON", http.StatusBadRequest)
+		return
+	}
+
+	result, err := h.db.Exec("INSERT INTO students (student_id, name, grade, phone, parent_phone) VALUES ($1, $2, $3, $4, $5)",
+		student.StudentID, student.Name, student.Grade, student.Phone, student.ParentPhone)
+	if err != nil {
+		log.Println("학생 등록 실패:", err)
+		http.Error(w, "Failed to create student", http.StatusInternalServerError)
+		return
+	}
+
+	affected, err := result.RowsAffected()
+	if err != nil {
+		log.Println("학생 등록 실패:", err)
+		http.Error(w, "Failed to create student", http.StatusInternalServerError)
+		return
+	}
+	if affected == 0 {
+		http.Error(w, "Student not found", http.StatusNotFound)
+		return
+	}
+
+	writeJSON(w, http.StatusCreated, map[string]string{"message": "Student created"})
+}
+
+func (h *Handler) UpdateStudentByID(w http.ResponseWriter, r *http.Request) {
+	id := getIDFromPath(r.URL.Path)
+	if id == "" {
+		http.Error(w, "Missing student ID", http.StatusBadRequest)
+		return
+	}
+
+	var student models.Student
+	if err := json.NewDecoder(r.Body).Decode(&student); err != nil {
+		http.Error(w, "Invalid JSON", http.StatusBadRequest)
+		return
+	}
+
+	result, err := h.db.Exec("UPDATE students SET name = $1, grade = $2, phone = $3, parent_phone = $4 WHERE student_id = $5",
+		student.Name, student.Grade, student.Phone, student.ParentPhone, id)
+	if err != nil {
+		log.Println("학생 수정 실패:", err)
+		http.Error(w, "Failed to update student", http.StatusInternalServerError)
+		return
+	}
+	affected, err := result.RowsAffected()
+	if err != nil {
+		log.Println("학생 수정 실패:", err)
+		http.Error(w, "Failed to update student", http.StatusInternalServerError)
+		return
+	}
+	if affected == 0 {
+		http.Error(w, "Student not found", http.StatusNotFound)
+		return
+	}
+
+	writeJSON(w, http.StatusOK, map[string]string{"message": "Student updated"})
+}
+
+func (h *Handler) DeleteStudentByID(w http.ResponseWriter, r *http.Request) {
+	id := getIDFromPath(r.URL.Path)
+	if id == "" {
+		http.Error(w, "Missing student ID", http.StatusBadRequest)
+		return
+	}
+
+	_, err := h.db.Exec("DELETE FROM students WHERE student_id = $1", id)
+	if err != nil {
+		log.Println("학생 삭제 실패:", err)
+		http.Error(w, "Failed to delete student", http.StatusInternalServerError)
+		return
+	}
+
+	writeJSON(w, http.StatusOK, map[string]string{"message": "Student deleted"})
+}
+
+func (h *Handler) GetStudentByID(w http.ResponseWriter, r *http.Request) {
+	id := getIDFromPath(r.URL.Path)
+	if id == "" {
+		http.Error(w, "Missing student ID", http.StatusBadRequest)
+		return
+	}
+
+	var student models.Student
+	row := h.db.QueryRow("SELECT * FROM students WHERE student_id = $1", id)
+	err := row.Scan(&student.StudentID, &student.Name, &student.Grade, &student.Phone, &student.ParentPhone)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			http.Error(w, "Student not found", http.StatusNotFound)
+		} else {
+			log.Println("학생 조회 실패:", err)
+			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+		}
+		return
+	}
+
+	writeJSON(w, http.StatusOK, student)
+}
